Udacity Meme Generator/requirements.txt
click==7.1.2
Flask==1.1.2
lxml==4.5.1
MarkupSafe==1.1.1
numpy==1.18.5
pandas==1.0.4
Pillow==7.1.2
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
requests==2.23.0
urllib3==1.25.9
Xpdfreader==4.0.3